WORKER_SERVER/requirements.txt
apache-airflow==3.1.6
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
pandas==2.3.3
numpy==2.2.6
selenium==4.40.0
webdriver-manager==4.0.2
psycopg2-binary==2.9.11
langchain==1.2.6
langchain-google-vertexai==3.2.1
langchain-postgres==0.0.16
fastapi==0.117.1
uvicorn==0.40.0
openai>=1.0.0
boto3>=1.26.0